backend seeder data default table data and  company backend api fix and frontednt implementatios

+++ backendApi/database/migrations/2023_08_17_103013_update_expance_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-	    Schema::table('expenses', function($table) {
-		    $table->double('refundable_amount',10,2)->default(0.00)->after('amount');
-		    $table->double('refunded_amount',10,2)->default(0.00)->after('refundable_amount');
-	    });
+        if (Schema::hasTable('expenses')) {
+            Schema::table('expenses', function (Blueprint $table) {
+                $table->double('refundable_amount', 10, 2)->default(0.00)->after('amount');
+                $table->double('refunded_amount', 10, 2)->default(0.00)->after('refundable_amount');
+            });
+        }
     }
 
     /**
@@ -22,9 +24,15 @@ return new class extends Migration
      */
     public function down(): void
     {
-	    Schema::table('expenses', function($table) {
-		    $table->dropColumn('refundable_amount');
-		    $table->dropColumn('refunded_amount');
-	    });
+        if (Schema::hasTable('expenses')) {
+            Schema::table('expenses', function (Blueprint $table) {
+                if (Schema::hasColumn('expenses', 'refundable_amount')) {
+                    $table->dropColumn('refundable_amount');
+                }
+                if (Schema::hasColumn('expenses', 'refunded_amount')) {
+                    $table->dropColumn('refunded_amount');
+                }
+            });
+        }
     }
 };
