+++ backendApi/database/migrations/2023_11_15_224112_add_column_on_debts_and_lends_table.php
@@ -11,23 +11,31 @@ return new class extends Migration
      */
 	public function up(): void
 	{
-		Schema::table('lends', function($table) {
-			$table->string('note')->after('amount')->default('null');
-		});
-		Schema::table('borrows', function($table) {
-			$table->string('note')->after('amount')->default('null');
-		});
+		if (Schema::hasTable('lends')) {
+			Schema::table('lends', function (Blueprint $table) {
+				$table->string('note')->after('amount')->default('null');
+			});
+		}
+		if (Schema::hasTable('borrows')) {
+			Schema::table('borrows', function (Blueprint $table) {
+				$table->string('note')->after('amount')->default('null');
+			});
+		}
 	}
 
 	/**
 	 * Reverse the migrations.
 	 */
 	public function down(): void {
-		Schema::table( 'lends', function ( Blueprint $table ) {
-			$table->dropColumn( 'note' );
-		} );
-		Schema::table( 'borrows', function ( Blueprint $table ) {
-			$table->dropColumn( 'note' );
-		} );
+		if (Schema::hasTable('lends') && Schema::hasColumn('lends', 'note')) {
+			Schema::table('lends', function (Blueprint $table) {
+				$table->dropColumn('note');
+			});
+		}
+		if (Schema::hasTable('borrows') && Schema::hasColumn('borrows', 'note')) {
+			Schema::table('borrows', function (Blueprint $table) {
+				$table->dropColumn('note');
+			});
+		}
 	}
 };
